requirements.txt
pyqrcode
pypng
colorama
pytube
apify_client
Pillow
googlesearch-python
pyshorteners